backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# Redis & Workers
redis>=5.0.0
rq>=1.15.0

# OIDC / JWT Authentication
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Production WSGI Server
gunicorn>=21.2.0

# Observability
prometheus-client>=0.19.0

# OIDC / JWT Authentication
PyJWT>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0

